pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eda-prototype"
version = "0.1.0"
description = "Executable Document Architecture — LLM-powered AOT compiler that synthesizes documents into deterministic, executable Python APIs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "EDA Research" },
]
keywords = ["llm", "document-compilation", "rag-alternative", "aot-compiler", "knowledge-extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "google-genai>=1.0.0",
    "PyMuPDF>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.30.0"]
rag-baseline = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
]
all = [
    "eda-prototype[openai,anthropic,rag-baseline]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "deepeval>=1.0.0",
]

[project.scripts]
eda = "eda.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/eda"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests that require LLM API calls (deselect with '-m \"not slow\"')",
    "benchmark: marks benchmark tests (deselect with '-m \"not benchmark\"')",
]
addopts = "-m 'not slow and not benchmark'"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
